dreamshot-multi/requirements.txt
python-telegram-bot[aiohttp]==20.7
aiohttp
requests
Pillow